Added Scoring Logic to Player, moved Deck logic to Dealer. Removed Deck and Hand Classes. Added Test classes for player and dealer.
