+++ src/com/javaninjas/blackjack/service/Player.java
@@ -17,7 +17,31 @@ public class Player {
     }
 
     //Business Method
-    //TODO: add scoring logic
+
+    public int scoreHand() {
+        // initialize aces to zero so we can do special scoring on aces
+        int aces = 0;
+        // maintains a local copy of score in the method
+        int score = 0;
+        for (Cards card : hand) {
+            if (card.getValue() == 0) {
+                aces++;
+            }
+            score += card.getValue();
+        }
+
+        if (aces != 0) {
+            for (int i = 0; i < aces; i++) {
+                if (getScore() + 11 < 21) {
+                    score += 11;
+                } else {
+                    score += 1;
+                }
+            }
+        }
+        return score;
+    }
+
     public void addCard(Cards card) {
         hand.add(card);
     }
