--- src/com/javaninjas/blackjack/service/PlayerFactory.java
@@ -1,7 +1,0 @@
-package com.javaninjas.blackjack.service;
-
-public class PlayerFactory {
-    public static PlayerFactory createPlayer(String name) {
-
-    }
-}
