+++ test/com/javaninjas/blackjack/service/DealerTest.java
@@ -1,4 +1,4 @@
 package com.javaninjas.blackjack.service;
 
-public class HandTest {
+class DealerTest {
 }
