--- test/com/javaninjas/blackjack/service/DeckTest.java
@@ -1,4 +1,0 @@
-package com.javaninjas.blackjack.service;
-
-public class DeckTest {
-}
