+++ test/com/javaninjas/blackjack/service/PlayerTest.java
@@ -1,4 +1,4 @@
 package com.javaninjas.blackjack.service;
 
-class TableTest {
+class PlayerTest {
 }
